// UI/src/app/Models/Crm/ApplicantInfo.ts
export interface IApplicantInfo {
  bookingId: number;
  firstApplicantName: string;
  firstApplicantGender: string;
  firstApplicantDateOfBirth: any;
  firstApplicantEmail: string;
  firstApplicantAlternateEmail: string;
  firstApplicantPhoneNumber: string;
  firstApplicantAlternatePhoneNumber: string;
  firstApplicantAadharNumber: string;
  firstApplicantPanNumber: string;
  firstApplicantSalutation: string;
  firstApplicantGuardianName: string;
  firstApplicantMaritalStatus: string;
  firstApplicantNationality: string;
  firstApplicantPassportNumber: string;
  firstApplicantProfession: string;
  firstApplicantDesignation: string;
  secondApplicantName: string;
  secondApplicantGender: string;
  secondApplicantDateOfBirth: any;
  secondApplicantEmail: string;
  secondApplicantAlternateEmail: string;
  secondApplicantPhoneNumber: string;
  secondApplicantAlternatePhoneNumber: string;
  secondApplicantAadharNumber: string;
  secondApplicantPanNumber: string;
  secondApplicantSalutation: string;
  secondApplicantGuardianName: string;
  secondApplicantMaritalStatus: string;
  secondApplicantNationality: string;
  secondApplicantPassportNumber: string;
  secondApplicantProfession: string;
  secondApplicantDesignation: string;
  basicPrice: number;
  finalPrice: number;
  projectId: number;
  levelId: number;
  bookedById: number;
  paymentPlanId: number;
  blockId: number;
  unitId: number;
  status: string;
  firstApplicantAddress1: string;
  firstApplicantAddress2: string;
  firstApplicantCity: string;
  firstApplicantPincode: string;
  firstApplicantState: string;
  secondApplicantAddress1: string;
  secondApplicantAddress2: string;
  secondApplicantCity: string;
  secondApplicantPincode: string;
  secondApplicantState: string;
  unitTypeName: string;
  basePrice: number;
  totalCharges: number;
  crmUserId: number;
  receivedAmount: number;
  pendingAmount: number;
  description: string;
  userId: number;
  gst: number;
  isMailSent: string;
  firstApplicantRelationToParent: string;
  firstApplicantParentOrSpouseSalutation: string;
  secondApplicantRelationToParent: string;
  secondApplicantParentOrSpouseSalutation: string;
  firstApplicantAnniversaryDate: string;
  secondApplicantAnniversaryDate: string;
  firstApplicantParentOrSpouse: string;
  secondApplicantParentOrSpouse: string;
  secondApplicantPanCard: string;
  secondApplicantAadharCard: string;
  firstApplicantPanCard: string;
  firstApplicantAadharCard: string;
  sourceName: string;
  subSourceName: string;
  preSalesId: number;
  thirdApplicantName: string;
  thirdApplicantGender: string;
  thirdApplicantDateOfBirth: string;
  thirdApplicantEmail: string;
  thirdApplicantAlternateEmail: string;
  thirdApplicantPhoneNumber: string;
  thirdApplicantAlternatePhoneNumber: string;
  thirdApplicantAadharNumber: string;
  thirdApplicantPanNumber: string;
  thirdApplicantSalutation: string;
  thirdApplicantGuardianName: string;
  thirdApplicantMaritalStatus: string;
  thirdApplicantNationality: string;
  thirdApplicantPassportNumber: string;
  thirdApplicantProfession: string;
  thirdApplicantDesignation: string;
  thirdApplicantPanCard: string;
  thirdApplicantAadharCard: string;
  thirdApplicantParentOrSpouse: string;
  thirdApplicantAnniversaryDate: string;
  thirdApplicantRelationToParent: string;
  thirdApplicantParentOrSpouseSalutation: string;
  thirdApplicantAddress1: string;
  thirdApplicantAddress2: string;
  thirdApplicantCity: string;
  thirdApplicantPincode: string;
  thirdApplicantState: string;
  receivedTds: number;
  isCredentialMailSent: string;
  firstAndSecondApplicantBothResidingatSameAddress: string;
  receivedPayment: number;
  applicantBankId: number;
  cancelBookingFormUrl: string;
  currentStageOwner: string;
  cancelRemarks: string;
  cancelledBookingDate: any;
  sanctionLetter: string;
}

export class ApplicantInfo implements IApplicantInfo {
  bookingId: number = 0;
  firstApplicantName: string = '';
  firstApplicantGender: string = '';
  firstApplicantDateOfBirth: any;
  firstApplicantEmail: string = '';
  firstApplicantAlternateEmail: string = '';
  firstApplicantPhoneNumber: string = '';
  firstApplicantAlternatePhoneNumber: string = '';
  firstApplicantAadharNumber: string = '';
  firstApplicantPanNumber: string = '';
  firstApplicantSalutation: string = '';
  firstApplicantGuardianName: string = '';
  firstApplicantMaritalStatus: string = '';
  firstApplicantNationality: string = '';
  firstApplicantPassportNumber: string = '';
  firstApplicantProfession: string = '';
  firstApplicantDesignation: string = '';
  secondApplicantName: string = '';
  secondApplicantGender: string = '';
  secondApplicantDateOfBirth: any;
  secondApplicantEmail: string = '';
  secondApplicantAlternateEmail: string = '';
  secondApplicantPhoneNumber: string = '';
  secondApplicantAlternatePhoneNumber: string = '';
  secondApplicantAadharNumber: string = '';
  secondApplicantPanNumber: string = '';
  secondApplicantSalutation: string = '';
  secondApplicantGuardianName: string = '';
  secondApplicantMaritalStatus: string = '';
  secondApplicantNationality: string = '';
  secondApplicantPassportNumber: string = '';
  secondApplicantProfession: string = '';
  secondApplicantDesignation: string = '';
  basicPrice: number = 0;
  finalPrice: number = 0;
  projectId: number = 0;
  levelId: number = 0;
  bookedById: number = 0;
  paymentPlanId: number = 0;
  blockId: number = 0;
  unitId: number = 0;
  status: string = '';
  firstApplicantAddress1: string = '';
  firstApplicantAddress2: string = '';
  firstApplicantCity: string = '';
  firstApplicantPincode: string = '';
  firstApplicantState: string = '';
  secondApplicantAddress1: string = '';
  secondApplicantAddress2: string = '';
  secondApplicantCity: string = '';
  secondApplicantPincode: string = '';
  secondApplicantState: string = '';
  unitTypeName: string = '';
  basePrice: number = 0;
  totalCharges: number = 0;
  crmUserId: number = 0;
  receivedAmount: number = 0;
  pendingAmount: number = 0;
  description: string = '';
  userId: number = 0;
  gst: number = 0;
  isMailSent: string = '';
  firstApplicantRelationToParent: string = '';
  firstApplicantParentOrSpouseSalutation: string = '';
  secondApplicantRelationToParent: string = '';
  secondApplicantParentOrSpouseSalutation: string = '';
  firstApplicantAnniversaryDate: string = '';
  secondApplicantAnniversaryDate: string = '';
  firstApplicantParentOrSpouse: string = '';
  secondApplicantParentOrSpouse: string = '';
  secondApplicantPanCard: string = '';
  secondApplicantAadharCard: string = '';
  firstApplicantPanCard: string = '';
  firstApplicantAadharCard: string = '';
  sourceName: string = '';
  subSourceName: string = '';
  preSalesId: number = 0;
  thirdApplicantName: string = '';
  thirdApplicantGender: string = '';
  thirdApplicantDateOfBirth: string = '';
  thirdApplicantEmail: string = '';
  thirdApplicantAlternateEmail: string = '';
  thirdApplicantPhoneNumber: string = '';
  thirdApplicantAlternatePhoneNumber: string = '';
  thirdApplicantAadharNumber: string = '';
  thirdApplicantPanNumber: string = '';
  thirdApplicantSalutation: string = '';
  thirdApplicantGuardianName: string = '';
  thirdApplicantMaritalStatus: string = '';
  thirdApplicantNationality: string = '';
  thirdApplicantPassportNumber: string = '';
  thirdApplicantProfession: string = '';
  thirdApplicantDesignation: string = '';
  thirdApplicantPanCard: string = '';
  thirdApplicantAadharCard: string = '';
  thirdApplicantParentOrSpouse: string = '';
  thirdApplicantAnniversaryDate: string = '';
  thirdApplicantRelationToParent: string = '';
  thirdApplicantParentOrSpouseSalutation: string = '';
  thirdApplicantAddress1: string = '';
  thirdApplicantAddress2: string = '';
  thirdApplicantCity: string = '';
  thirdApplicantPincode: string = '';
  thirdApplicantState: string = '';
  receivedTds: number = 0.0;
  isCredentialMailSent: string = '';
  firstAndSecondApplicantBothResidingatSameAddress: string = '';
  receivedPayment: number = 0;
  applicantBankId: number = 0;
  cancelBookingFormUrl: string = '';
  currentStageOwner: string = '';
  cancelRemarks: string = '';
  cancelledBookingDate: any;
  sanctionLetter: string = '';
}